requirements.txt
beautifulsoup4>=4.6.0
boto3>=1.7.37
click>=6.7
docb>=0.9.2
PyYAML>=3.12
requests>=2.19.0
sammy>=0.2.3